feature(Tomoe) refactor
