--- Tomoe/Core/src/extensions/administration/config.json
@@ -1,6 +1,0 @@
-{
-  "name": "administration",
-  "displayName": "Admin",
-  "injectables": {
-  }
-}
