+++ Tomoe/Core/src/extensions/admissions/src/server/classes/Hacker.js
