+++ Tomoe/Core/src/extensions/admissions/src/server/handlers/hacker.js
@@ -1,4 +1,4 @@
-import { Hacker } from '../collections'
+import { Hacker } from '../classes'
 import { operatorSearch, respond, transformData, setSearchParam } from '../../../importable';
 import Joi from 'joi';
 import Boom from 'boom';
