+++ Tomoe/Core/src/extensions/volunteers/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "Tomoe-Admin",
+  "name": "template",
   "version": "1.0.0",
-  "description": "Admin utility for Tomoe",
+  "description": "A template for Tomoe Extensions",
   "main": " ",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
