+++ Tomoe/Core/src/extensions/volunteers/server/classes/index.js
@@ -1,0 +1,1 @@
+export * from './Volunteer'
