+++ Tomoe/Core/src/extensions/volunteers/server/routes/index.js
@@ -1,0 +1,7 @@
+import volunteerRoutes from './volunteer';
+
+const routes = [
+  ...volunteerRoutes,
+];
+
+export default routes;
