+++ Tomoe/Core/src/extensions/volunteers/server/validators/index.js
@@ -1,0 +1,1 @@
+export * from './volunteer'
