+++ Tomoe/Core/src/server/exportable.js
@@ -1,4 +1,4 @@
 // content that can be utilized by extensions
 
-export * from './src/collections';
+export * from './src/classes';
 export * from './src/util';
