+++ Tomoe/Core/src/server/src/classes/generic/Event.js
@@ -1,0 +1,19 @@
+import uuidv4 from 'uuid/v4'
+
+export class Event {
+  constructor({ id = uuidv4(), date = null }){
+    this.id = id;
+    this.date = date;
+  }
+
+  static getCollection(){
+    return db.collection('event');
+  }
+
+  _get(){
+    return {
+      id: this.id,
+      date: this.date,
+    }
+  }
+}
