+++ Tomoe/Core/src/server/src/classes/generic/index.js
@@ -1,0 +1,2 @@
+export * from './User';
+export * from './Event';
