+++ Tomoe/Core/src/server/src/classes/index.js
@@ -1,0 +1,2 @@
+export * from './generic';
+export * from './strict';
