+++ Tomoe/Core/src/server/src/classes/strict/Admin.js
@@ -1,4 +1,4 @@
-import { User } from './';
+import { User } from '../';
 
 export class Admin extends User{
   constructor(data = {}){
