+++ Tomoe/Core/src/server/src/classes/strict/index.js
@@ -1,0 +1,1 @@
+export * from './Admin';
