--- Tomoe/Core/src/server/src/collections/index.js
@@ -1,3 +1,0 @@
-export * from './User';
-export * from './Admin';
-export * from './Volunteer';
