+++ Tomoe/Core/src/server/src/handlers/index.js
@@ -1,2 +1,1 @@
 export * from './admin'
-export * from './volunteer'
