+++ Tomoe/Core/src/server/src/routes/index.js
@@ -1,11 +1,9 @@
 import adminRoutes from './admin';
-import volunteerRoutes from './volunteer';
 
 import { getExtensionContent } from '../util';
 
 const routes = [
   ...adminRoutes,
-  ...volunteerRoutes,
   ...getExtensionContent('routes'),
 ];
